Set up DynamoDB interface (#3)
This sets up a database interface for the rest of the project to use, abstracting away Dynamo.

There's an integration test that verifies the interface works against Amazon's local dynamodb, as well as a mock that's suitable for use in unit tests of other parts of the codebase. We run the same smoketest against both to verify they behave equivalently in the basic codepaths we've got in the database interface.