Early removal detection (#7)
This adds a check for certificates which are removed from a CRL early.

What is early? The certificate must appear in at least one CRL while
already expired.

This uses the unauthenticated /get/cert boulder API to fetch
certificates by their serial number.

Ideally, some of this code will be part of Boulder. We can decide how
much in a boulder PR.