Add cert churner (#10)
This adds the code for the "Cert churner". It is not yet a lambda, but
ships a standalone cli tool meant for use in dev. In a future change,
we'll refactor out the common code for the lambda.

This uses a new acme account every request, which isn't great. That will
be fixed in another future PR.

There's not a ton of test coverage over the ACME code, which a future
integration test will cover, but I don't want to block on that right now
so we can write a unit test that covers a more complete system later on.