Add CRL Checker (#11)
This adds the core of the CRL checker. The checker is what actually
examines CRLs.

The command-line program is nearly what will drive this as a lambda once
s3 event handling code is added, but is suitable for use in dev right
now. It successfully lints current CRLs and alerts on early removal from
our old CRLs during the previous incident.

Some fixups to other parts of code to make this work:
- Return aws.EndpointNotFound error in resolver so the sdk handles it
properly
- Don't make any API request to delete 0 certs
- Log every 100 certs fetched