Update README.md diagram
Having the CRL checker alert if a serial number is missing from the CRLs
isn't sufficient because it runs on S3 events, so would fail to alert if
no CRLs at all were uploaded and thus never ran.

This changes that check to exist in the churner, which runs on a timer.