Check for serials missing from CRLs (#12)
This adds a check to the churner which looks for missing CRLs. If a
certificate hasn't appeared in a CRL within REVOKE_DEADLINE, then an
error is thrown.

I hadn't originally planned on putting this logic in the CRL Churner,
but I think it makes sense: The churner runs on a timer, which is what
we need, and it's already going to be connecting to DynamoDB. I've
updated the diagram in README.md to match.

With this change, this project is now largely feature-complete, and
successfully runs in dev. Some remaining work remains to make it run as
a lambda, along with the terraform module to configure it.

This also removes RandDomains from the public API of the churner, since
it isn't intended for use outside that module.