Add a script and GH action to produce builds (#24)
This adds a shell script to create the .zip files for the checker and
churner lambdas.

It adds a set of github actions:

try-release.yml just runs the build but doesn't do anything with it.
This is runs on PRs and when they're pushed to main.

release.yml runs the build and then uploads the artifacts to Github
whenever a tag is pushed starting with "v" (which is how normal go
release are tagged). I anticipate replacing this with uploading to S3 in
a future PR, as we need them in S3 for lambda to deploy from.