Check for CRLDistributionPoint in certificates.
If a certificate has a CRLDistributionPoint, it will be stored in the database
in a new "DP" column.

At check time, each CRL must have exactly one IssuingDistributionPoint.
When a certificate is seen, its CRLDistributionPoint will be compared to the
IssuingDistributionPoint of the CRL it was seen on. If they differ, error.